+++ sandbox/nedit/gui/app/Pin.hpp
@@ -12,7 +12,7 @@ struct Pin
 {
     Pin(ax::NodeEditor::PinId init_id,
         ax::NodeEditor::PinKind init_kind,
-        std::uint32_t init_type,
+        std::uint64_t init_type,
         std::string_view init_label,
         const FlowIcon& init_icon);
 
@@ -27,7 +27,7 @@ struct Pin
     std::unordered_set<std::uint64_t> links;
     ax::NodeEditor::PinId id;
     ax::NodeEditor::PinKind kind;
-    std::uint32_t type;
+    std::uint64_t type;
     std::string_view label;
     const FlowIcon& icon;
 };
